+++ test/gc_util.dart
@@ -135,25 +135,3 @@ class Log {
   static printOff(String name) => Logger(name).clearListeners();
 }
 
-class X {
-  final int value;
-  X(this.value);
-
-  @override
-  int get hashCode => value.hashCode;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) || (other is X && value == other.value);
-
-  @override
-  String toString() => 'X$value';
-}
-
-class Y {
-  String value;
-  Y(int value) : value = '$value';
-
-  @override
-  String toString() => 'Y$value';
-}