+++ test/weak_list_gc_2_test.dart
@@ -7,6 +7,8 @@ import 'dart:math';
 import 'package:test/test.dart';
 import 'package:weak/weak.dart';
 import 'gc_util.dart';
+import 'list_util.dart';
+import 'data_util.dart';
 
 /// IMPORTANT: set VM options
 ///
@@ -16,211 +18,6 @@ import 'gc_util.dart';
 void main() async {
   final vmService = await VmServiceUtil.create();
 
-  void Function(X? elm) expectEach(List<X?> expected) {
-    int j = 0;
-    return (elm) => expect(elm, equals(expected[j++]));
-  }
-
-  void expectGetRange(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      for (int e = s; e <= len; ++e) {
-        expect(list.getRange(s, e), equals(expected.getRange(s, e)),
-            reason: '($s, $e)');
-      }
-    }
-    expect(() => list.getRange(len + 1, len + 1), throwsRangeError);
-  }
-
-  void expectAny(WeakList<X> list, List<X?> expected) {
-    expect(list.any((e) => e == null), equals(expected.any((e) => e == null)));
-    final len = expected.length;
-    for (int s = 0; s < len; ++s) {
-      final x = expected[s];
-      expect(list.any((e) => e == x), equals(expected.any((e) => e == x)));
-    }
-  }
-
-  void expectAsMap(WeakList<X> list, List<X?> expected) {
-    expect(list.asMap().keys, orderedEquals(expected.asMap().keys));
-    expect(list.asMap().values, orderedEquals(expected.asMap().values));
-  }
-
-  void expectContains(WeakList<X> list, List<X?> expected) {
-    expect(list.contains(null), equals(expected.contains(null)));
-    final len = expected.length;
-    for (int s = 0; s < len; ++s) {
-      final x = expected[s];
-      expect(list.any((e) => e == x), equals(expected.any((e) => e == x)));
-    }
-  }
-
-  void expectElementAt(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    expect(() => list.elementAt(-1), throwsRangeError);
-    expect(() => list.elementAt(len), throwsRangeError);
-    for (int j = 0; j < len; ++j) {
-      expect(list.elementAt(j), equals(expected.elementAt(j)), reason: '$j');
-    }
-  }
-
-  void expectEvery(WeakList<X> list, List<X?> expected) {
-    expect(
-        list.every((e) => e == null), equals(expected.every((e) => e == null)));
-    final len = expected.length;
-    for (int s = 0; s < len; ++s) {
-      final x = expected[s];
-      expect(list.every((e) => e == x), equals(expected.every((e) => e == x)));
-    }
-    for (int s = 0; s < len; ++s) {
-      expect(list.every((e) => e == null || e.value < len),
-          equals(expected.every((e) => e == null || e.value < len)));
-    }
-  }
-
-  void expectIndexOf(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.indexOf(null, s), equals(expected.indexOf(null, s)));
-    }
-    if (len > 0) {
-      final elm = expected[Random().nextInt(len)];
-      for (int s = 0; s <= len; ++s) {
-        expect(list.indexOf(elm, s), equals(expected.indexOf(elm, s)));
-      }
-    }
-  }
-
-  void expectLastIndexOf(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.lastIndexOf(null, s), equals(expected.lastIndexOf(null, s)));
-    }
-    if (len > 0) {
-      final elm = expected[Random().nextInt(len)];
-      for (int s = 0; s <= len; ++s) {
-        expect(list.lastIndexOf(elm, s), equals(expected.lastIndexOf(elm, s)),
-            reason: 'start: $s, elm: $elm');
-      }
-    }
-  }
-
-  void expectIndexWhere(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.indexWhere((elm) => elm == null, s),
-          equals(expected.indexWhere((elm) => elm == null, s)));
-    }
-    if (len > 0) {
-      final elm0 = expected[Random().nextInt(len)];
-      for (int s = 0; s <= len; ++s) {
-        expect(list.indexWhere((elm) => elm == elm0, s),
-            equals(expected.indexWhere((elm) => elm == elm0, s)));
-      }
-    }
-  }
-
-  void expectLastIndexWhere(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.lastIndexWhere((elm) => elm == null, s),
-          equals(expected.lastIndexWhere((elm) => elm == null, s)));
-    }
-    if (len > 0) {
-      final elm0 = expected[Random().nextInt(len)];
-      for (int s = 0; s <= len; ++s) {
-        expect(list.lastIndexWhere((elm) => elm == elm0, s),
-            equals(expected.lastIndexWhere((elm) => elm == elm0, s)));
-      }
-    }
-  }
-
-  void expectWhere(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    expect(list.where((elm) => elm == null),
-        equals(expected.where((elm) => elm == null)));
-    for (int s = 0; s <= len; ++s) {
-      expect(list.where((elm) => elm != null && elm.value < s),
-          equals(expected.where((elm) => elm != null && elm.value < s)));
-    }
-  }
-
-  void expectFirstWhere(WeakList<X> list, List<X?> expected, X? elseX) {
-    final len = expected.length;
-    expect(list.firstWhere((elm) => elm == null, orElse: () => elseX),
-        equals(expected.firstWhere((elm) => elm == null, orElse: () => elseX)));
-    for (int s = 0; s < len; ++s) {
-      final x = expected[s];
-      expect(list.firstWhere((elm) => elm == x, orElse: () => null),
-          equals(expected.firstWhere((elm) => elm == x, orElse: () => null)));
-    }
-  }
-
-  void expectLastWhere(WeakList<X> list, List<X?> expected, X? elseX) {
-    final len = expected.length;
-    expect(list.lastWhere((elm) => elm == null, orElse: () => elseX),
-        equals(expected.lastWhere((elm) => elm == null, orElse: () => elseX)));
-    for (int s = 0; s < len; ++s) {
-      final x = expected[s];
-      expect(list.lastWhere((elm) => elm == x, orElse: () => null),
-          equals(expected.lastWhere((elm) => elm == x, orElse: () => null)));
-    }
-  }
-
-  void expectSingleWhere(WeakList<X> list, List<X?> expected, X? elseX) {
-    final len = expected.length;
-    expect(
-        list.singleWhere((elm) => elm == null, orElse: () => elseX),
-        equals(
-            expected.singleWhere((elm) => elm == null, orElse: () => elseX)));
-    for (int s = 0; s < len; ++s) {
-      final x = expected[s];
-      expect(list.singleWhere((elm) => elm == x, orElse: () => null),
-          equals(expected.singleWhere((elm) => elm == x, orElse: () => null)));
-    }
-  }
-
-  void expectSkip(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.skip(s), equals(expected.skip(s)));
-    }
-  }
-
-  void expectSkipWhile(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.skipWhile((elm) => elm == null),
-          equals(expected.skipWhile((elm) => elm == null)));
-    }
-    if (len > 0) {
-      for (int s = 0; s <= len; ++s) {
-        expect(list.skipWhile((elm) => elm == null || elm.value < s),
-            equals(expected.skipWhile((elm) => elm == null || elm.value < s)));
-      }
-    }
-  }
-
-  void expectTake(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.take(s), equals(expected.take(s)));
-    }
-  }
-
-  void expectTakeWhile(WeakList<X> list, List<X?> expected) {
-    final len = expected.length;
-    for (int s = 0; s <= len; ++s) {
-      expect(list.takeWhile((elm) => elm == null),
-          equals(expected.takeWhile((elm) => elm == null)));
-    }
-    if (len > 0) {
-      for (int s = 0; s <= len; ++s) {
-        expect(list.takeWhile((elm) => elm == null || elm.value < s),
-            equals(expected.takeWhile((elm) => elm == null || elm.value < s)));
-      }
-    }
-  }
 
   group("empty or one element list ", () {
     late WeakList<X> list;
@@ -257,6 +54,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectLastIndexOf(list, expected);
       expectIndexWhere(list, expected);
@@ -319,6 +120,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectIndexWhere(list, expected);
       expectLastIndexOf(list, expected);
@@ -394,6 +199,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectIndexWhere(list, expected);
       expectLastIndexOf(list, expected);
@@ -456,6 +265,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectIndexWhere(list, expected);
       expectLastIndexOf(list, expected);
@@ -533,6 +346,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectLastIndexOf(list, expected);
       expectIndexWhere(list, expected);
@@ -619,6 +436,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectLastIndexOf(list, expected);
       expectIndexWhere(list, expected);
@@ -841,6 +662,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectLastIndexOf(list, expected);
       expectIndexWhere(list, expected);
@@ -934,6 +759,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectLastIndexOf(list, expected);
       expectIndexWhere(list, expected);
@@ -1173,6 +1002,44 @@ void main() async {
       }
       for (int s = 0; s < len; ++s) {
         list.insertAll(s, extra);
+        strongList.insertAll(s, extra);
+        expected.insertAll(s, expectExtra);
+        expect(list, orderedEquals(expected));
+      }
+      list.validElementCount();
+      extra = <X?>[];
+      final nullified = <X?>{};
+      for (int j = 0; j < expected.length; ++j) {
+        final elm = expected[j];
+        if (j.isOdd) {
+          nullified.add(elm);
+          strongList[j] = null;
+          expected[j] = null;
+        }
+      }
+      for (int j = 0; j < expected.length; ++j) {
+        if (nullified.contains(expected[j])) {
+          strongList[j] = null;
+          expected[j] = null;
+        }
+      }
+
+      await vmService.gc();
+
+      expect(list, isNotEmpty);
+      expect(list.length, equals(expected.length));
+
+      expect(list, orderedEquals(expected), reason: '$list');
+      list.validElementCount();
+    });
+
+    test('insertNulls()', () async {
+      final len = expected.length;
+      const extraCount = count ~/ 2;
+      List<X?> extra = List<X?>.filled(extraCount, null);
+      final expectExtra = List<X?>.filled(extraCount, null);
+      for (int s = 0; s < len; ++s) {
+        list.insertNulls(s, extraCount);
         strongList.insertAll(s, extra);
         expected.insertAll(s, expectExtra);
         expect(list, orderedEquals(expected));
@@ -1462,6 +1329,18 @@ void main() async {
       }
     });
 
+    test('insertNulls()', () {
+      final len = expected.length;
+      const extraCount = count ~/ 2;
+      final extra = List<X?>.filled(extraCount, null);
+      for (int s = 0; s < len; ++s) {
+        list.insertNulls(s, extraCount);
+        list.validElementCount();
+        expected.insertAll(s, extra);
+        expect(list, orderedEquals(expected));
+      }
+    });
+
     test('remove()', () {
       final len = expected.length;
       for (int j = 0; j < len; ++j) {
@@ -1647,6 +1526,10 @@ void main() async {
       list.forEach(expectEach(expected));
 
       expectGetRange(list, expected);
+      expectNonNulls(list, expected);
+      expectNonNullsReversed(list, expected);
+      expectGetNonNullRange(list, expected);
+      expectGetNonNullRangeReversed(list, expected);
       expectIndexOf(list, expected);
       expectLastIndexOf(list, expected);
       expectIndexWhere(list, expected);
@@ -1763,6 +1646,18 @@ void main() async {
       }
       for (int s = 0; s < len; ++s) {
         list.insertAll(s, extra);
+        list.validElementCount();
+        expected.insertAll(s, extra);
+        expect(list, orderedEquals(expected));
+      }
+    });
+
+    test('insertNulls()', () {
+      final len = expected.length;
+      const extraCount = count ~/ 2;
+      final extra = List.filled(extraCount, null);
+      for (int s = 0; s < len; ++s) {
+        list.insertNulls(s, extraCount);
         list.validElementCount();
         expected.insertAll(s, extra);
         expect(list, orderedEquals(expected));
@@ -1900,19 +1795,4 @@ void main() async {
       }
     });
   });
-}
-
-class X {
-  final int value;
-  X(this.value);
-
-  @override
-  int get hashCode => value.hashCode;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) || (other is X && value == other.value);
-
-  @override
-  String toString() => 'X$value';
 }