+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+- add [WeakList.nonNulls] for the case that no need to iterate nulls (for better performance)
+- and [WeakList.nonNullsReversed] for the case that no need to iterate nulls reversely (for better performance)
+- add [WeakList.getNonNullRange] for the case that no need to iterate nulls in a range
+- add [WeakList.getNonNullRangeReversed] for the case that no need to iterate nulls reversely in a range
+- tune [WeakList.where] about non-null iteration for better performance 
+
 ## 1.0.0+2
 
 - run dart fix on lints 5.1.1 (package:lints/recommended.yaml)